% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invregCRE.R
\name{compATE_CRE}
\alias{compATE_CRE}
\title{Inverse Regression for Completely Randomized Experiments with Multiple Outcomes}
\usage{
compATE_CRE(y, z, x, tau, alpha)
}
\arguments{
\item{y}{A numeric matrix of observed multiple outcomes, where rows represent individuals and
columns represent distinct outcomes.}

\item{z}{A numeric vector of assigned treatments, where each element corresponds to an individual.}

\item{x}{A numeric vector or matrix of covariates to be included in the inverse regression.
Rows correspond to individuals and columns to covariates.}

\item{tau}{A logical value indicating whether the vector of true marginal treatment effects is zero.}

\item{alpha}{A numeric value specifying the confidence level for the confidence interval.}
}
\value{
A list of class \code{'res'} containing the following components:
\describe{
\item{ATEc}{Estimated average treatment effect (ATE) for the composite outcome.}
\item{SE}{Standard error of the ATE estimator for the composite outcome.}
\item{CI}{Confidence interval for the ATE of the composite outcome.}
\item{lambda}{Estimated weights of the asymptotic distribution of the ATE estimator
for the composite outcome when \code{tau = 0}.}
}
}
\description{
Implements an inverse regression approach to estimating the average treatment effect (ATE)
for multiple outcomes in completely randomized experiments.
}
\details{
The \code{compATE_CRE()} function computes the ATE estimator, its standard error, and confidence interval
for a composite outcome based on the inverse regression framework. The function supports two settings:
estimation without covariates and estimation with covariates included in the inverse regression.
In both settings, the function reports results when the vector of true marginal treatment effects
is either zero or nonzero.
}
\examples{
n <- 1000
L <- 9
beta <- matrix(rep(0.1, L), nrow = n, ncol = L, byrow = TRUE)
K <- 5
eta <- matrix(0.01, nrow = K, ncol = L)
p <- 0.5
z <- rbinom(n, size = 1, prob = p)
x <- matrix(rnorm(n * K), ncol = K, nrow = n)
epsilon <- matrix(rnorm(n * L), ncol = L, nrow = n)
y <- matrix(z, nrow = n, ncol = L) * beta + x \%*\% eta + epsilon

compATE_CRE(y = y, z = z, x = x, tau = "Nonzero", alpha = 0.05)

}
