% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invregOS.R
\name{compATE_OS}
\alias{compATE_OS}
\title{Inverse Regression for Observational Studies with Multiple Outcomes}
\usage{
compATE_OS(y, z, x, xinInvreg, tau, alpha)
}
\arguments{
\item{y}{A numeric matrix of observed multiple outcomes, where rows represent individuals and
columns represent distinct outcomes.}

\item{z}{A numeric vector of assigned treatments, where each element corresponds to an individual.}

\item{x}{A numeric vector or matrix of pretreatment covariates. Rows correspond to individuals and columns to covariates.}

\item{xinInvreg}{A logical value indicating whether the covariates are included in the weighted inverse regression}

\item{tau}{A logical value indicating whether the vector of true marginal treatment effects is zero.}

\item{alpha}{A numeric value specifying the confidence level for the confidence interval.}
}
\value{
A list of class \code{'res'} containing the following components:
\describe{
\item{ATEc}{Estimated average treatment effect (ATE) for the composite outcome.}
\item{SE}{Standard error of the ATE estimator for the composite outcome.}
\item{CI}{Confidence interval for the ATE for the composite outcome.}
\item{lambda}{Estimated weights for the asymptotic distribution of the ATE estimator for the composite outcome
when \code{tau = 0}.}
}
}
\description{
Implements a weighted inverse regression approach to estimating the average treatment effect (ATE)
for multiple outcomes in observational studies.
}
\details{
The \code{compATE_OS()} function computes the ATE estimator, its standard error, and confidence interval
for a composite outcome based on the weighted inverse regression framework. The function supports two settings:
estimation without covariates and estimation with covariates included in the weighted inverse regression.
In both settings, the function reports results when the vector of true marginal treatment effects
is either zero or nonzero.
}
\examples{
n <- 1000
L <- 9
beta <- matrix(rep(0.1, L), nrow = n, ncol = L, byrow = TRUE)
K <- 5
eta <- matrix(0.01, nrow = K, ncol = L)
alpha <- rep(0.1, K+1)
x <- matrix(rnorm(n * K), ncol = K, nrow = n)
xring.alpha <- cbind(rep(1,n), x) \%*\% matrix(alpha, ncol=1)
px <- exp(xring.alpha)/(1+exp(xring.alpha))
z <- apply(matrix(px, ncol=1), MARGIN=1, function(u) rbinom(1, size=1, prob=u))
epsilon <- matrix(rnorm(n*L), ncol=L, nrow=n)
y <- matrix(z, nrow=n, ncol=L)*beta + x \%*\% eta + epsilon

compATE_OS(y=y, z=z, x=x, xinInvreg=FALSE, tau="Nonzero", alpha=0.05)

}
