% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invregSRER.R
\name{compATE_SRER}
\alias{compATE_SRER}
\title{Inverse Regression for Stratified Randomized Experiments with Multiple Outcomes}
\usage{
compATE_SRER(y, z, c, cf, x, r, tauSR, alpha)
}
\arguments{
\item{y}{A numeric matrix of observed multiple outcomes, where rows represent individuals and
columns represent distinct outcomes.}

\item{z}{A numeric vector of assigned treatments, where each element corresponds to an individual.}

\item{c}{A numeric vector of observed stratum indicators #'}

\item{cf}{A factor of possible strata indicators}

\item{x}{A numeric vector or matrix of covariates to be included in the inverse regression.
Rows correspond to individuals and columns to covariates.}

\item{r}{An adjusting coefficient in the covariate-adjusted ATE estimators.}

\item{tauSR}{A logical value indicating whether the vector of variance-weighted stratum-specific
treatment effects is equal zero.}

\item{alpha}{A numeric value specifying the confidence level for the confidence interval.}
}
\value{
A list of class \code{'res'} containing the following components:
\describe{
\item{ATEc}{Estimated average treatment effect (ATE) for the composite outcome.}
\item{SE}{Standard error of the ATE estimator for the composite outcome.}
\item{CI}{Confidence interval for the ATE of the composite outcome.}
\item{lambda}{Estimated weights for the asymptotic distribution of the ATE estimator for the composite outcome
when \code{tauSR = 0}.}
}
}
\description{
Implements an inverse regression approach to estimating the average treatment effect (ATE)
for multiple outcomes in stratified randomized experiments based on the regression strategy.
}
\details{
The \code{compATE_SRE()} function computes the ATE estimator, its standard error, and confidence interval
for a composite outcome based on the inverse regression framework. The function supports two settings:
estimation without covariates and estimation with covariates included in the inverse regression.
In both settings, the function reports results when the vector of variance-weighted stratum-specific
treatment effects is either zero or nonzero.
}
\examples{
n <- 1000
L <- 9
beta <- rep(0.1, L)
K <- 5
eta <- matrix(0.01, nrow = K, ncol = L)
S <- 3
cf <- as.factor(c(1:S))
p <- c(0.3, 0.5, 0.7)
n.stra <- rmultinom(n = 1, size = n, prob = c(0.3, 0.3, 0.4))
z <- NULL
x <- NULL
y <- NULL
c <- NULL
for(s in 1:S){
ns <- n.stra[s,1]
tempz <- rbinom(ns, size=1, prob=p[s])
z <- c(z, tempz)
tempx <- matrix(rnorm(ns*K), ncol=K, nrow=ns)
x <- rbind(x, tempx)
epsilons <- matrix(rnorm(ns*L), ncol=L, nrow=ns)
tempy <- matrix(tempz, nrow=ns, ncol=L)*matrix(beta, nrow=ns, ncol=L, byrow=T) + tempx\%*\%eta + epsilons
y <- rbind(y, tempy)
tempc <- rep(s, ns)
c <- c(c, tempc)
}
compATE_SRER(y=y, z=z, c=c, cf=cf, x=NULL, r=-1, tauSR="Nonzero", alpha=0.05)

}
