% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DavisApprox.R
\name{q_mixchisq_davies}
\alias{q_mixchisq_davies}
\title{Davies method}
\usage{
q_mixchisq_davies(p, lambda)
}
\arguments{
\item{p}{A numeric value whose sample quantile is wanted}

\item{lambda}{The weights of the weighted sum of chi-squared distributions
to be evaluated}
}
\value{
A value is returned.
}
\description{
Quantiles of a weighted sum of chi-squared distributions
using Davies's approximation method.
}
\details{
The \code{q_mixchisq_davies()} function computes the quantiles of
weighted sum of chi-squared distributions using Davies's approximation method.
See details in R fuction davies() in the package CompQuadForm.
}
