% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTZApprox.R
\name{q_mixchisq_liu}
\alias{q_mixchisq_liu}
\title{Liu-Tang-Zhang method}
\usage{
q_mixchisq_liu(p, lambda)
}
\arguments{
\item{p}{A numeric value whose sample quantile is wanted}

\item{lambda}{The weights of the weighted sum of chi-squared distributions
to be evaluated}
}
\value{
A value is returned.
}
\description{
Quantiles of a weighted sum of chi-squared distributions
using Liu-Tang-Zhang approximation method.
}
\details{
The \code{q_mixchisq_liu()} function computes the quantiles of
weighted sum of chi-squared distributions using Liu-Tang-Zhang approximation method.
See details in R fuction liu() in the package CompQuadForm.
}
