% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighMoment.R
\name{Swab}
\alias{Swab}
\title{Weighted sample moments}
\usage{
Swab(a, b, w)
}
\arguments{
\item{a}{A matrix of \eqn{n\times L_1}}

\item{b}{A matrix of \eqn{n\times L_2}}

\item{w}{A numeric vector of length \eqn{n}}
}
\value{
A value is returned.
}
\description{
Weighted moments of two matrix.
}
\details{
The \code{Swab()} function computes the weighted moments of two matrix
with the same number of rows: \eqn{n^{-1}\sum_{i=1}^{n}w_ia_ib_i^{'}}
}
